Create launcher.py
This was created in response to this program by the ModuloProject which aims to make a command line app launcher. The method they were using to get data from launchers seemed sub-optimal so I reworked it to be a bit more efficient. Their original version is found at: http://git.io/vWJcm